Replace trialsTable
