+++ dashboard/src/components/trialsTable.tsx
@@ -1,104 +1,226 @@
-import React, { FC, useState } from "react"
-import { makeStyles } from "@material-ui/core/styles"
-import {
-  TableFooter,
-  TablePagination,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Paper,
-} from "@material-ui/core"
-import TablePaginationActions from "@material-ui/core/TablePagination/TablePaginationActions"
+import React, { FC } from "react"
+import { createStyles, makeStyles, Theme } from "@material-ui/core/styles"
+import Table from "@material-ui/core/Table"
+import TableBody from "@material-ui/core/TableBody"
+import TableCell from "@material-ui/core/TableCell"
+import TableContainer from "@material-ui/core/TableContainer"
+import TableHead from "@material-ui/core/TableHead"
+import TablePagination from "@material-ui/core/TablePagination"
+import TableRow from "@material-ui/core/TableRow"
+import TableSortLabel from "@material-ui/core/TableSortLabel"
 
-const useStyles = makeStyles({
-  table: {
-    minWidth: 650,
-  },
-})
+function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
+  if (b[orderBy] < a[orderBy]) {
+    return -1
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1
+  }
+  return 0
+}
+
+type Order = "asc" | "desc"
+
+function getComparator<Key extends keyof any>(
+  order: Order,
+  orderBy: Key
+): (
+  a: { [key in Key]: number | string },
+  b: { [key in Key]: number | string }
+) => number {
+  return order === "desc"
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy)
+}
+
+function stableSort<T>(array: T[], comparator: (a: T, b: T) => number) {
+  const stabilizedThis = array.map((el, index) => [el, index] as [T, number])
+  stabilizedThis.sort((a, b) => {
+    const order = comparator(a[0], b[0])
+    if (order !== 0) return order
+    return a[1] - b[1]
+  })
+  return stabilizedThis.map((el) => el[0])
+}
+
+interface HeadCell {
+  id: keyof Trial
+  label: string
+}
+
+const headCells: HeadCell[] = [
+  { id: "trial_id", label: "Trial ID" },
+  { id: "number", label: "Number" },
+  { id: "state", label: "State" },
+  { id: "value", label: "Value" },
+  { id: "params", label: "Params" },
+]
+
+interface TrialsTableProps {
+  classes: ReturnType<typeof useStyles>
+  onRequestSort: (
+    event: React.MouseEvent<unknown>,
+    property: keyof Trial
+  ) => void
+  order: Order
+  orderBy: string
+}
+
+function TrialsTableHead(props: TrialsTableProps) {
+  const { classes, order, orderBy, onRequestSort } = props
+  const createSortHandler = (property: keyof Trial) => (
+    event: React.MouseEvent<unknown>
+  ) => {
+    onRequestSort(event, property)
+  }
+
+  return (
+    <TableHead>
+      <TableRow>
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id}
+            sortDirection={orderBy === headCell.id ? order : false}
+          >
+            <TableSortLabel
+              active={orderBy === headCell.id}
+              direction={orderBy === headCell.id ? order : "asc"}
+              onClick={createSortHandler(headCell.id)}
+            >
+              {headCell.label}
+              {orderBy === headCell.id ? (
+                <span className={classes.visuallyHidden}>
+                  {order === "desc" ? "sorted descending" : "sorted ascending"}
+                </span>
+              ) : null}
+            </TableSortLabel>
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  )
+}
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {
+      width: "100%",
+    },
+    table: {
+      minWidth: 750,
+    },
+    visuallyHidden: {
+      border: 0,
+      clip: "rect(0 0 0 0)",
+      height: 1,
+      margin: -1,
+      overflow: "hidden",
+      padding: 0,
+      position: "absolute",
+      top: 20,
+      width: 1,
+    },
+  })
+)
 
 export const TrialsTable: FC<{
-  trials: Trial[]
-}> = ({ trials = [] }) => {
+  rows: Trial[]
+}> = ({ rows = [] }) => {
   const classes = useStyles()
-  const [page, setPage] = useState<number>(0)
-  const [rowsPerPage, setRowsPerPage] = useState<number>(10)
+  const [order, setOrder] = React.useState<Order>("asc")
+  const [orderBy, setOrderBy] = React.useState<keyof Trial>("trial_id")
+  const [page, setPage] = React.useState(0)
+  const [rowsPerPage, setRowsPerPage] = React.useState(10)
 
-  const emptyRows =
-    rowsPerPage - Math.min(rowsPerPage, trials.length - page * rowsPerPage)
-  const handleChangePage = (
-    event: React.MouseEvent<HTMLButtonElement> | null,
-    newPage: number
+  const handleRequestSort = (
+    event: React.MouseEvent<unknown>,
+    property: keyof Trial
   ) => {
+    const isAsc = orderBy === property && order === "asc"
+    setOrder(isAsc ? "desc" : "asc")
+    setOrderBy(property)
+  }
+
+  const handleChangePage = (event: unknown, newPage: number) => {
     setPage(newPage)
   }
 
   const handleChangeRowsPerPage = (
-    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
+    event: React.ChangeEvent<HTMLInputElement>
   ) => {
-    const value = parseInt(event.target.value, 10)
-    setRowsPerPage(value)
+    setRowsPerPage(parseInt(event.target.value, 10))
     setPage(0)
   }
 
+  const emptyRows =
+    rowsPerPage - Math.min(rowsPerPage, rows.length - page * rowsPerPage)
+
+  const sortedRows = stableSort(rows, getComparator(order, orderBy))
+  const paginateRows =
+    rowsPerPage > 0
+      ? sortedRows.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+      : sortedRows
+
   return (
-    <TableContainer component={Paper}>
-      <Table className={classes.table} size="small" aria-label="trials table">
-        <TableHead>
-          <TableRow>
-            <TableCell>Trial ID</TableCell>
-            <TableCell>Number</TableCell>
-            <TableCell>State</TableCell>
-            <TableCell>Value</TableCell>
-            <TableCell>Params</TableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {(rowsPerPage > 0
-            ? trials.slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-            : trials
-          ).map((t) => {
-            return (
-              <TableRow key={t.trial_id}>
-                <TableCell component="th" scope="row">
-                  {t.trial_id}
-                </TableCell>
-                <TableCell>{t.number}</TableCell>
-                <TableCell>{t.state.toString()}</TableCell>
-                <TableCell>{t.value}</TableCell>
-                <TableCell>
-                  {t.params.map((p) => p.name + ": " + p.value).join(", ")}
-                </TableCell>
+    <div className={classes.root}>
+      <TableContainer>
+        <Table
+          className={classes.table}
+          aria-labelledby="tableTitle"
+          size="small"
+          aria-label="enhanced table"
+        >
+          <TrialsTableHead
+            classes={classes}
+            order={order}
+            orderBy={orderBy}
+            onRequestSort={handleRequestSort}
+          />
+          <TableBody>
+            {paginateRows.map((row, index) => {
+              const labelId = `enhanced-table-checkbox-${index}`
+
+              return (
+                <TableRow
+                  hover
+                  role="checkbox"
+                  tabIndex={-1}
+                  key={row.trial_id}
+                >
+                  <TableCell
+                    component="th"
+                    id={labelId}
+                    scope="row"
+                    align="right"
+                  >
+                    {row.trial_id}
+                  </TableCell>
+                  <TableCell align="right">{row.number}</TableCell>
+                  <TableCell align="right">{row.state.toString()}</TableCell>
+                  <TableCell align="right">{row.value}</TableCell>
+                  <TableCell align="left">
+                    {row.params.map((p) => p.name + ": " + p.value).join(", ")}
+                  </TableCell>
+                </TableRow>
+              )
+            })}
+            {emptyRows > 0 && (
+              <TableRow style={{ height: 33 * emptyRows }}>
+                <TableCell colSpan={6} />
               </TableRow>
-            )
-          })}
-          {emptyRows > 0 && (
-            <TableRow style={{ height: 53 * emptyRows }}>
-              <TableCell colSpan={6} />
-            </TableRow>
-          )}
-        </TableBody>
-        <TableFooter>
-          <TableRow>
-            <TablePagination
-              rowsPerPageOptions={[10, 50, 100, { label: "All", value: -1 }]}
-              colSpan={3}
-              count={trials.length}
-              rowsPerPage={rowsPerPage}
-              page={page}
-              SelectProps={{
-                inputProps: { "aria-label": "rows per page" },
-                native: true,
-              }}
-              onChangePage={handleChangePage}
-              onChangeRowsPerPage={handleChangeRowsPerPage}
-              ActionsComponent={TablePaginationActions}
-            />
-          </TableRow>
-        </TableFooter>
-      </Table>
-    </TableContainer>
+            )}
+          </TableBody>
+        </Table>
+      </TableContainer>
+      <TablePagination
+        rowsPerPageOptions={[10, 50, 100, { label: "All", value: -1 }]}
+        component="div"
+        count={rows.length}
+        rowsPerPage={rowsPerPage}
+        page={page}
+        onChangePage={handleChangePage}
+        onChangeRowsPerPage={handleChangeRowsPerPage}
+      />
+    </div>
   )
 }
